Revert "Update audioCacheManager.js"
This reverts commit 555c398a5bd7445d901b1e75f7eb351bb2024b5b.

+++ function/audioCacheManager.js
@@ -218,7 +218,7 @@ class AudioCacheManager {
             isClosed = true;
 
             try {
-                if (fs.existsSync(tempFile)) {
+                if (fs.existsSync(tempFile) && totalBytesWritten >= 3840) {
                     fs.renameSync(tempFile, targetFile);
                     this.cacheMap.set(videoId, {
                         filePath: targetFile,
